+++ lib/src/web3_signers_base.dart
@@ -11,12 +11,11 @@ import 'package:passkeys/types.dart';
 import 'package:web3dart/crypto.dart';
 import 'package:web3dart/web3dart.dart';
 
-import 'package:web3_signers/src/vendor/vendor.dart';
+import 'package:web3_signers/src/vendor/vendor.dart' show TupleExtension;
 
 import 'interfaces/interfaces.dart';
 import 'utils/utils.dart';
 
 part 'signers/eoa_wallet.dart';
-part 'signers/hardware_signer.dart';
 part 'signers/passkey_signer.dart';
 part 'signers/private_key_signer.dart';